Refactored HPAExtension tests
